+++ components/Navbar.tsx
@@ -56,20 +56,24 @@ export const Navbar = () => {
               <Disclosure.Panel className="sm:hidden">
                 <div className="space-y-1 pt-2 pb-4">
                   {/* Current: "bg-indigo-50 border-indigo-500 text-indigo-700", Default: "border-transparent text-gray-500 hover:bg-gray-50 hover:border-gray-300 hover:text-gray-700" */}
-                  <Disclosure.Button
-                    as="a"
-                    href="#"
-                    className="block border-l-4 border-indigo-500 bg-indigo-50 py-2 pl-3 pr-4 text-base font-medium text-indigo-700"
-                  >
-                    Data Dashboard
-                  </Disclosure.Button>
-                  <Disclosure.Button
-                    as="a"
-                    href="#"
-                    className="block border-l-4 border-transparent py-2 pl-3 pr-4 text-base font-medium text-gray-500 hover:border-gray-300 hover:bg-gray-50 hover:text-gray-700"
-                  >
-                    Slow Zones
-                  </Disclosure.Button>
+                  <ActiveLink href="/" activeClassName="border-white ">
+                    <Disclosure.Button
+                      as="a"
+                      href="#"
+                      className="block border-l-4 border-transparent py-2 pl-3 pr-4 text-base font-medium text-white hover:border-gray-300 hover:bg-gray-50 hover:text-gray-700"
+                    >
+                      Data Dashboard
+                    </Disclosure.Button>
+                  </ActiveLink>
+                  <ActiveLink href="/slowzones" activeClassName="border-white">
+                    <Disclosure.Button
+                      as="a"
+                      href="#"
+                      className="block border-l-4 border-transparent py-2 pl-3 pr-4 font-medium text-white  hover:border-white hover:bg-gray-50 hover:text-gray-700"
+                    >
+                      Slow Zones
+                    </Disclosure.Button>
+                  </ActiveLink>
                 </div>
               </Disclosure.Panel>
             </>
