+++ components/dashboard/charts/Legend.tsx
