+++ components/dashboard/charts/styles/Legend.module.css
