+++ components/slowzones/charts/LineSegments.tsx
@@ -1,0 +1,3 @@
+export const LineSegments = () => {
+  return <div></div>;
+};
