+++ components/slowzones/charts/TotalSlowTime.tsx
@@ -1,0 +1,3 @@
+export const TotalSlowTime = () => {
+  return <div></div>;
+};
